requirements.txt
flask==3.0.3
web3==6.20.3
python-dotenv==1.0.1
py-solc-x==2.0.3
